rename: 모달 컴포넌트 위치 변경에 따른 import 수정
