+++ src/page/MyPage/EditApplication.tsx
@@ -9,7 +9,7 @@ import { CheckBoxSelect } from '@/components/common/CheckBox/CheckBoxSelect';
 import axios from 'axios';
 import { ReactComponent as ModalClose } from '@/assets/icons/signup/ModalClose.svg';
 import { ReactComponent as CloseButtonX } from '@/assets/icons/mypage/CloseButton.svg';
-import Modal from '@/components/common/Modal';
+import Modal from '@/components/common/Modal/Modal';
 import { Area } from '@/data/Area';
 
 interface Area {
